requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0  # For async SQLite support

# Security dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email dependencies
emails==0.6
jinja2==3.1.2

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Image processing and AI dependencies
torch==2.1.0+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.0+cpu --index-url https://download.pytorch.org/whl/cpu
timm==0.9.16
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.25.2
fastai==2.7.12  # For BI-RADS breast cancer classification

# TensorFlow/Keras dependencies for stroke classification
tensorflow==2.15.0

# Cloud storage dependencies
cloudinary==1.36.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2